requirements.txt
certifi==2020.12.5
chardet==4.0.0
httplib2==0.19.1
idna==2.10
oauth2==1.9.0.post1
progressbar2==3.53.1
pyparsing==2.4.7
python-utils==2.5.6
requests==2.25.1
sanction==0.4.1
six==1.16.0
urllib3==1.26.6